[CV9] Scaling (& Changelog)
View elements no longer take a number for constant dimensions; instead, they accept a function which will return a value. Further internal optimisations to be made, and possibly allowance for providing element-specific scalings. Currently, scaling is provided based on a single constant, and relative to 1080p.

Scaling currently does not work for text.

Also tweaked changelog formatting because why not :)