Improve Movable & Resizable frames
Resizing is now limited by the size of the frame's contents. To achieve this, `MovableResizableFrame`now delegates sizing to a `SizeControl` object, which is placed as the `MovableFrame`'s first child. `MovableResizableFrame` still needs to be `MovableFrame`'s parent so it can "move" `MovableFrame` around when resizing takes place. 

Offsets are also now measured to top-left corner of the frame, as users expect windows to have a constant top-left corner when auto-resizing. 

Also, `MovableFrame` now ensures at least 5 pixels of its drag handle are on-screen, to prevent situations where it cannot be dragged.

Offsets now can be obtained with `MovableFrame:CurrentOffset()`.

Other:
- Rename `MovableFrame`'s `vStack` to `zStack` (coz it's not a `VerticalStack`, it's a `StackInPlace`)