`table.repeating`, `table.ifilter`,`table.joinArrays`
`table.repeating` returns an array with a set number of values, with a generation function that takes the array index of the value to generate and returns the value to be stored at that index.

`table.ifilter` returns a new array with all elements of a given array where the provided filter returns true. The filter is given the index and value for each entry in the array. Order is preserved.

`table.joinArrays`returns an array containing all elements in the provided arrays, in reverse of the order that was provided.