Replace `IsAboveWatcher` with `HighestResponderAtPoint` and loops
Adds `responder:ContainsPoint(x, y)` which will allow overriding behaviour e.g. for custom shapes
Adds `responder.mouseIsOver`, which is guaranteed to correctly reflect the mouseover state of all responders handling `events.mouseOver` when `responder:MouseEnter()` or `responder:MouseLeave()` are called, even if those responders are in a different element. Previously, `responder:MouseLeave()` and `responder:MouseEnter()` were called as the state updates were made, such that any logic depending on the mouseover state of other responders (in both the same element & other element) could only be guaranteed to be operating on a correct value if the mouseover responder was known to have higher priority in the responder chain (either being in a higher priority element or by being known to be in a higher priority responder).

This should allow more correct handling of e.g. `Menu`'s submenus